working on bug, cleaned up bookings component. Created utils for dateFormatter
